feat: add events for old arch
